+++ app/models/sku.rb
@@ -17,12 +17,14 @@ class Sku < ActiveRecord::Base
   belongs_to :account
   attr_accessible :num_iid, :properties, :properties_name, :quantity, :sku_id,
        :taobao_product_id, :account_id,:product_id, :sku_properties_attributes,:code
-  belongs_to :product,:inverse_of => :skus
+  belongs_to :product,:include => :category ,:inverse_of => :skus
   has_many :sku_bindings, dependent: :destroy
   has_many :stock_products
   has_many :taobao_skus, through: :sku_bindings
   has_many :sku_properties,  :dependent=>:destroy
   accepts_nested_attributes_for :sku_properties
+
+  has_one :category, through: :product
 
 
   after_save  :migrate_taobao_sku_props